requirements.txt
streamlit
openai
qdrant-client
python-dotenv
pydub
streamlit-audiorecorder
